+++ demo/index.js
@@ -1,7 +1,7 @@
 import CodeMirror from 'codemirror'
 import debounce from 'lodash.debounce'
 import { jsxslack } from '../src/index'
-import example from './example'
+import * as _examples from './example'
 import schema from './schema'
 
 import 'codemirror/mode/javascript/javascript'
@@ -17,10 +17,19 @@ const jsx = document.getElementById('jsx')
 const json = document.getElementById('json')
 const error = document.getElementById('error')
 const errorDetails = document.getElementById('errorDetails')
+const examplesSelect = document.getElementById('examples')
 const previewBtn = document.getElementById('preview')
 
 json.addEventListener('click', () => json.select())
 
+const parseHash = (hash = window.location.hash) => {
+  if (!hash.toString().startsWith('#')) return undefined
+  return decodeURIComponent(hash.toString().slice(1))
+}
+
+const examples = Object.assign(Object.create(null), _examples)
+
+// CodeMirror
 const completeAfter = (cm, pred) => {
   if (!pred || pred())
     setTimeout(() => {
@@ -62,20 +71,36 @@ const jsxEditor = CodeMirror(jsx, {
   indentUnit: 2,
   lineWrapping: true,
   mode: 'jsx',
-  value: example,
+  value: (() => {
+    const hash = parseHash()
+
+    if (examples[hash]) {
+      examplesSelect.value = hash
+      return examples[hash]
+    }
+
+    return examples.blockKit
+  })(),
 })
 
 const convert = () => {
   try {
     const output = jsxslack([jsxEditor.getValue()])
-
     json.value = JSON.stringify(output, null, '  ')
-    previewBtn.setAttribute(
-      'href',
-      `https://api.slack.com/tools/block-kit-builder?blocks=${encodeURIComponent(
-        JSON.stringify(output)
-      )}`
-    )
+
+    if (Array.isArray(output)) {
+      previewBtn.removeAttribute('tabindex')
+      previewBtn.setAttribute(
+        'href',
+        `https://api.slack.com/tools/block-kit-builder?blocks=${encodeURIComponent(
+          JSON.stringify(output)
+        )}`
+      )
+      previewBtn.classList.remove('disabled')
+    } else {
+      previewBtn.setAttribute('tabindex', -1)
+      previewBtn.classList.add('disabled')
+    }
 
     error.classList.add('hide')
   } catch (e) {
@@ -87,6 +112,18 @@ const convert = () => {
   }
 }
 
-jsxEditor.on('change', debounce(convert, 600))
+const onChangeEditor = debounce(convert, 600)
+jsxEditor.on('change', onChangeEditor)
 
 convert()
+
+examplesSelect.addEventListener('change', () => {
+  if (examplesSelect.value) {
+    jsxEditor.off('change', onChangeEditor)
+    setTimeout(() => jsxEditor.on('change', onChangeEditor), 0)
+
+    window.location.hash = `#${examplesSelect.value}`
+    jsxEditor.setValue(examples[examplesSelect.value])
+    convert()
+  }
+})